+++ client-pulsefit/src/features/pattern-engine/index.ts
@@ -1,0 +1,4 @@
+export * from './types'
+export * from './detectors'
+export * from './recommendation-builder'
+export * from './insight-prioritizer'